requirements.txt
python-telegram-bot[webhooks]==21.6
yt-dlp>=2025.10.01
yt-dlp-youtube-oauth2
flask
